requirements.txt
streamlit==1.24.0
google-generativeai==0.3.1
pandas==1.5.3
plotly==5.14.1
Pillow==9.5.0
python-dotenv

